+++ src/mechanics/card_logic.py
@@ -1,0 +1,8 @@
+class CardInfo:
+    def __init__(self,speed,action):
+        self.speed=speed
+        self.action = action
+    def get_speed(self):
+        return self.speed
+    def get_action(self):
+        return self.action