+++ Documentation/changelog.md
@@ -1,1 +1,3 @@
 ## Week 3
+- Made a general drawpile class to handle drawing and shufflig the deck
+- Tests for the drawpile have been added